dnfdaemon: Consistently use uint64_t in repository callbacks

+++ dnfdaemon-server/callbacks.hpp
@@ -85,7 +85,7 @@ public:
         long int timestamp) override;
 
 private:
-    double total = 0;
+    uint64_t total = 0;
 };
